% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-sf.R
\name{collate_extr_shapes}
\alias{collate_extr_shapes}
\title{Collate extraction shapes}
\usage{
collate_extr_shapes(sf, draw)
}
\arguments{
\item{sf}{loaded in-built sf}

\item{draw}{list of drawn leaflet features}
}
\value{
sf of input sf and draw features
}
\description{
Collate extraction shapes
}
