% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-sf.R
\name{drawFeature2sf}
\alias{drawFeature2sf}
\title{Convert a drawn leaflet feature to sf}
\usage{
drawFeature2sf(feature)
}
\arguments{
\item{feature}{drawn leaflet feature}
}
\value{
a simple feature object of the leaflet feature
}
\description{
Convert a drawn leaflet feature to sf
}
