% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-sf.R
\name{extr_sedmap_data}
\alias{extr_sedmap_data}
\title{Extract sedmap data}
\usage{
extr_sedmap_data(rst, sf, select_rst = NULL, select_sf = NULL,
  output = c("summaries", "csv", "raster"), fun = c("mean", "min",
  "max", "median", "sd"), sf_crs = FALSE, out_dir,
  rst_out_format = c("stack", "tiff"), select_sf_csv = FALSE,
  attributes, varnames)
}
\arguments{
\item{rst}{sedmaps data raster stack}

\item{sf}{extraction simple feature}

\item{select_rst}{raster layers to be extracted}

\item{select_sf}{sf layers to use for extraction}

\item{output}{character vector of extraction output formats}

\item{fun}{summary functions to apply for each data x extraction layer. \code{sd}
ignored for points}

\item{sf_crs}{whether output data should be converted to the sf crs}

\item{out_dir}{path to output directory}

\item{rst_out_format}{whether raster data should be return as a \code{.grd} raster
stack or individual geoTIFF layers}

\item{select_sf_csv}{whether to additionally include extraction sf metadata as in \code{.csv} format}

\item{attributes}{dataspice attributes table}

\item{varnames}{varnames list}
}
\value{
outputs written out to \code{out_dir}.
}
\description{
Extract sedmap data
}
