% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-panels.R
\name{fileName_meta}
\alias{fileName_meta}
\title{Get descriptive label}
\usage{
fileName_meta(fileName, return_field = "name")
}
\arguments{
\item{fileName}{the file for which metadata is to be returned}

\item{return_field}{character string. field to be returned for fileName}
}
\value{
a character string descriptive label of the variable
}
\description{
Create a more descriptive label, including units where appropriate, for a
variable from information in the metadata table.
}
