% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_basemap}
\alias{lflt_basemap}
\title{Plot leaflet basemap}
\usage{
lflt_basemap(rst, basemap = "Esri.OceanBasemap")
}
\arguments{
\item{rst}{sedMaps \code{rasterStack}}
}
\description{
Plot leaflet basemap
}
