% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_plot}
\alias{lflt_plot}
\title{Title}
\usage{
lflt_plot(rst, varname = NULL, basemap = "Esri.OceanBasemap",
  option = "A", label = NULL, opacity = 0.8)
}
\arguments{
\item{rst}{}

\item{varname}{}

\item{basemap}{}

\item{option}{}

\item{label}{}

\item{opacity}{}
}
\description{
Title
}
