% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_rst_selected}
\alias{lflt_rst_selected}
\title{Plot selected rst layer}
\usage{
lflt_rst_selected(rst, varname = NULL, option = "A", label = NULL,
  opacity = 0.8)
}
\arguments{
\item{rst}{sedMaps \code{rasterStack}}

\item{varname}{selected sedMaps \code{rasterStack} layer}

\item{option}{selected \code{viridis} palette option}

\item{label}{legend label. Defaults to \code{varname}}

\item{opacity}{layer opacity}
}
\description{
Plot selected rst layer reactively using leafletProxy
}
