% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_sf}
\alias{lflt_sf}
\title{Add sf vector selection layer}
\usage{
lflt_sf(map, sf = NULL, fillOpacity = 0.3, ...)
}
\arguments{
\item{map}{leaflet map object}

\item{sf}{sf object containing vector data}

\item{fillOpacity}{sf polygon fill colour}

\item{color}{sf polygon stroke colour}
}
\value{
a leaflet map object with sf vector data overlaid
}
\description{
Add sf vector selection layer
}
