% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_sf_selected}
\alias{lflt_sf_selected}
\title{Highlight selected sf polygons}
\usage{
lflt_sf_selected(sf, ids, pal_f = topo.colors(10), ...)
}
\arguments{
\item{sf}{sf object containing vector data}

\item{ids}{character vector of selected \code{id} ids}

\item{pal_f}{character vector of colors}

\item{...}{additional variables passed to \code{addPolylines()}.}
}
\description{
Highlight selected sf polygons
}
