% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataspice.R
\name{load_spice}
\alias{load_spice}
\title{Load `dataspice`` metadata tables}
\usage{
load_spice(metadata_path = here::here("data", "metadata"),
  table = NULL)
}
\arguments{
\item{metadata_path}{}

\item{table}{character string vector of names of valid \code{dataspice} metadata
tables. One of \code{"access"}, \code{"attributes"}, \code{"biblio"}, \code{"creators"})}
}
\value{
tibble loaded into the global environment for each valid \code{table}.
}
\description{
Load `dataspice`` metadata tables into the global environment
}
\examples{
\dontrun{
load_spice()
load_spice(table = "attributes")
}
}
