% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf-read.R
\name{ncdf_dimstack}
\alias{ncdf_dimstack}
\title{Stack bands from a third dimension into a single rasterStack}
\usage{
ncdf_dimstack(path, varname = NULL, dimension)
}
\arguments{
\item{path}{path to the ncdf file}

\item{varname}{single varname for which to extract bands}

\item{dimension}{name of dimension from which to extract bands. Check available dimensions with \code{ncdf_extra_dims()}.}
}
\value{
a stack of rasterLayers containing values for \code{varname} in each dimensional band.
}
\description{
Stack bands from a third dimension (e.g. time, depth) into a single rasterStack.
}
