% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf-read.R
\name{ncdf_stack}
\alias{ncdf_stack}
\title{Load a ncdf file into a rasterBrick}
\usage{
ncdf_stack(path)
}
\arguments{
\item{path}{path to ncdf file}
}
\value{
a rasterBrick, one layer for each varname in the ncdf.
}
\description{
Load a ncdf file into a rasterBrick
}
