% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf-utils.R
\name{ncdf_varnames}
\alias{ncdf_varnames}
\title{Extract metadata from ncdf object}
\usage{
ncdf_varnames(ncdf)
}
\arguments{
\item{ncdf}{ncdf object from \code{ncdf4::nc_open()}.}

\item{dimension}{the dimension from which to extracts information, See \code{ncdf_dim()}}
}
\value{
values returned vary according to \code{ncdf_*} function called:
\itemize{
\item \strong{\code{ncdf_varnames}} a vector of the varnames of each layer in the ncdf file.
\item \strong{\code{ncdf_ndims}} the number of dimensions in the ncdf.
\item \strong{\code{ncdf_dim}} a vector of ncdf dimension names
\item \strong{\code{ncdf_extra_dims}} a vector of dimension names additional to spatial lon/lat ($x$ & $y$) coordinates.
\item \strong{\code{ncdf_bands}} the bands of a specified dimension (pixel size).
\item \strong{\code{ncdf_band_units}} the units of the bands in a specified dimension.
}
}
\description{
Extract metadata from ncdf object
}
