% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-panels.R
\name{panel_layers}
\alias{panel_layers}
\title{Create layers panel}
\usage{
panel_layers(panel = c("sed", "dis"), mode = c("view", "extract"),
  varnames = varnames)
}
\arguments{
\item{panel}{}

\item{mode}{}
}
\description{
Create layers panel
}
