% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-panels.R
\name{select_button}
\alias{select_button}
\title{Title}
\usage{
select_button(panel = c("layers", "extract"), mode = c("select",
  "deselect"))
}
\arguments{
\item{panel}{}

\item{mode}{}
}
\description{
Title
}
