% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-utils.R
\name{update_selected_varnames}
\alias{update_selected_varnames}
\title{Update selected varnames during extract mode}
\usage{
update_selected_varnames(selected = NULL, new)
}
\arguments{
\item{selected}{last state of selected varnames}

\item{new}{new state of selected varname following last click.}
}
\value{
a vector of updated selected varnames. If \code{new} contains a new
\code{varname}, the \code{varname} is added to the end of \code{selected}. If a \code{varname}
has been removed in \code{new}, retuns \code{selected} with the \code{varname} removed.
}
\description{
Update selected varnames during extract mode
}
