% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_contour}
\alias{lflt_contour}
\title{add contour map derived from raster}
\usage{
lflt_contour(map, r, color = "white", weight = 0.5)
}
\arguments{
\item{map}{a leaflet map widget}

\item{r}{a raster to contour}

\item{color}{colour for the contour lines}

\item{weight}{weight of contour lines.}
}
\value{
a leaflet map widgets with contours from raster image added.
}
\description{
add contour map derived from raster
}
