% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-utils.R
\name{lflt_sf}
\alias{lflt_sf}
\title{Add sf vector selection layer}
\usage{
lflt_sf(map, sf, pal_f = topo.colors(10), ...)
}
\arguments{
\item{map}{leaflet map object}

\item{sf}{sf object containing vector data}

\item{pal_f}{character vector of colors}

\item{...}{additional variables passed to \code{addPolylines()}.}
}
\value{
a leaflet map object with sf vector data overlaid
}
\description{
Add sf vector selection layer
}
